@@ -16,11 +16,9 @@ limitations under the License.
 
 package config
 
-import (
-//"io/ioutil"
-//"os"
-//"testing"
-)
+// "io/ioutil"
+// "os"
+// "testing"
 
 var configStr = `
 {
@@ -50,7 +48,7 @@ var configStr = `
 
 //func TestLoadOK(t *testing.T) {
 //	content := []byte(configStr)
-//	tmpConfigFile, err := ioutil.TempFile(homeDir(), "kubewatch")
+//	tmpConfigFile, err := os.CreateTemp(homeDir(), "kubewatch")
 //	if err != nil {
 //		t.Fatalf("TestLoad(): %+v", err)
 //	}
