@@ -21,12 +21,13 @@ import (
 	"net/http"
 	"os"
 
+	_ "net/http/pprof"
+
 	"github.com/bitnami-labs/kubewatch/config"
 	c "github.com/bitnami-labs/kubewatch/pkg/client"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	_ "net/http/pprof"
 )
 
 var cfgFile string
@@ -46,9 +47,11 @@ supported webhooks:
  - slack
  - hipchat
  - mattermost
+ - msteams
  - flock
  - webhook
  - lark
+ - discord
 `,
 
 	Run: func(cmd *cobra.Command, args []string) {
